yeni model classı, crud işlemleri ve önyüz sayfaları eklendi

+++ src/main/java/com/ahmetalbayrak/service/BaseService.java
@@ -13,6 +13,6 @@ public interface BaseService<T>  {
     
     public void save(T entity);
     
-    public void delete(int id);
+    public void delete(Long id);
     
 }
